update in backend

+++ frontend/src/Project.js
@@ -5,10 +5,18 @@ import AddTask from './AddTask'
 import AddMember from './AddMember'
 import Nav from './Nav'
 
+
 function Project() {
     const [data,setData]=React.useState([])
     const [projectName,setProjectName]=React.useState("")
-    
+    const [members,setMembers]=React.useState([])
+    const [userEmail,setUserEmail]=React.useState("")
+    const [projectUserEmail,setProjectUserEmail]=React.useState("")
+    const [showComments,setShowComments]=React.useState(false)
+    const [comments,setComments]=React.useState([])
+    const [msg,setMsg]=React.useState("")
+    const [taskId,setTaskId]=React.useState("")
+    const [taskName,setTaskName]=React.useState("")
     const [flag,setFlag]=React.useState("")
     
 
@@ -39,6 +47,33 @@ function Project() {
                 setData(res.data.taskList)
                 setProjectName(res.data.name)
                 setFlag(res.flag)
+                setMembers(res.data.members)
+                setProjectUserEmail(res.data.userId.email)
+                
+            }
+            else{
+                if(res.message=='not getting token' || res.message=='token expired'){
+                    navigate('/login')
+                }
+                else{
+                    alert(res.message)
+                }
+            }
+            
+            
+        })
+    },[])
+    React.useEffect(()=>{
+        fetch("http://localhost:4000/getUserProjects",{
+            headers:{
+                'x-api-key':token
+            }
+        })
+        .then((result)=>result.json())
+        .then(res=>{
+            
+            if(res.status==true){
+                setUserEmail(res.data.email)
             }
             else{
                 if(res.message=='not getting token' || res.message=='token expired'){
@@ -54,7 +89,8 @@ function Project() {
     },[])
 
 
-    function dragToProgress(id){
+    function dragToProgress(id,assignTo){
+        if(assignTo!=userEmail && userEmail!=projectUserEmail)return alert("you can not dragged")
         let data={id:id,status:'InProgress'}
         fetch("http://localhost:4000/editTask",{
             method:"PUT",
@@ -76,7 +112,8 @@ function Project() {
         })
 
     }
-    function dragToCompleted(id){
+    function dragToCompleted(id,assignTo){
+        if(assignTo!=userEmail && userEmail!=projectUserEmail)return alert("you can not dragged")
         let data={id:id,status:'Completed'}
         fetch("http://localhost:4000/editTask",{
             method:"PUT",
@@ -98,7 +135,8 @@ function Project() {
         })
 
     }
-    function dragToDelete(id){
+    function dragToDelete(id,assignTo){
+        if(assignTo!=userEmail && userEmail!=projectUserEmail)return alert("you can not dragged")
         let data={id:id}
         fetch("http://localhost:4000/deleteTask",{
             method:"DELETE",
@@ -120,16 +158,104 @@ function Project() {
         })
 
     }
-
+    function getComments(taskId,name){
+        let data={"taskId":taskId}
+        setTaskId(taskId)
+        fetch("http://localhost:4000/getTaskById",{
+            method:"POST",
+            headers:{
+                'Content-type': 'application/json',
+                'x-api-key':token
+            },
+            body:JSON.stringify(data)
+        })
+        .then((result)=>result.json())
+        .then(res=>{
+            
+            if(res.status==true){
+                setShowComments(true)
+                setComments(res.data.comments)
+                setTaskName(name)
+            }
+            else{
+                if(res.message=='not getting token' || res.message=='token expired'){
+                    navigate('/login')
+                }
+                else{
+                    alert(res.message)
+                }
+            }
+            
+            
+        })
+    }
+    function addComment(){
+        let data={taskId:taskId,message:msg,userId:userId}
+        fetch("http://localhost:4000/createComment",{
+            method:"POST",
+            headers:{
+                'Content-type': 'application/json',
+                'x-api-key':token
+            },
+            body:JSON.stringify(data)
+        })
+        .then((result)=>result.json())
+        .then(res=>{
+            
+            if(res.status==true){
+                alert("comment added successfully")
+                window.location.reload(true)
+            }
+            else{
+                if(res.message=='not getting token' || res.message=='token expired'){
+                    navigate('/login')
+                }
+                else{
+                    alert(res.message)
+                }
+            }
+            
+            
+        })
+    }
+    function deleteComment(commentId){
+        let data={commentId:commentId,taskId:taskId}
+        fetch("http://localhost:4000/deleteComment",{
+            method:"PUT",
+            headers:{
+                'Content-type': 'application/json',
+                'x-api-key':token
+            },
+            body:JSON.stringify(data)
+        })
+        .then((result)=>result.json())
+        .then(res=>{
+            
+            if(res.status==true){
+                alert("comment deleted successfully")
+                window.location.reload(true)
+            }
+            else{
+                if(res.message=='not getting token' || res.message=='token expired'){
+                    navigate('/login')
+                }
+                else{
+                    alert(res.message)
+                }
+            }
+            
+            
+        })
+    }
 
   return (
     <>
-    <Nav name={userName}/>
+    <Nav name={userName} />
     <div className='nav2'>
         {
             flag?
                 <>
-                <AddTask projectId={params.projectId}/>
+                <AddTask projectId={params.projectId} members={members} userEmail={userEmail}/>
                 <AddMember projectId={params.projectId}/></>
             :
             <></>
@@ -139,7 +265,9 @@ function Project() {
     </div>
 
     <h1 style={{textAlign:"center"}}>project name: {projectName}</h1>
-    <div className='tasks'>
+     {
+        !showComments?
+        <div className='tasks'>
         <div className='Name'>
             <div><h3>Task to complete</h3></div>
             <div><h3>In-Progress</h3></div>
@@ -152,9 +280,10 @@ function Project() {
             {
                 data.map(x=>
                     x.status=="Todo" && x.isDeleted==false?
-                    <li draggable onDragEnd={(e)=>dragToProgress(x._id)}>
+                    <li draggable onDragEnd={(e)=>dragToProgress(x._id,x.assignTo)}>
                         <h4>{x.name}</h4><br/>
-                        {/* <p>{x.description}</p> */}
+                        <p>assignTo: {x.assignTo}</p>
+                        <a style={{cursor: "pointer"}} onClick={()=>{getComments(x._id,x.name)}}><p>comments: {x.totalComment}</p></a>
                     </li>:<></>
                 )
             }
@@ -163,9 +292,10 @@ function Project() {
             {
                 data.map(x=>
                     x.status=="InProgress" && x.isDeleted==false?
-                    <li  draggable onDragEnd={(e)=>dragToCompleted(x._id)} >
+                    <li  draggable onDragEnd={(e)=>dragToCompleted(x._id,x.assignTo)} >
                         <h4>{x.name}</h4><br/>
-                        {/* <p>{x.description}</p> */}
+                        <p>assignTo: {x.assignTo}</p>
+                        <a style={{cursor: "pointer"}} onClick={()=>{getComments(x._id,x.name)}}><p>comments: {x.totalComment}</p></a>
                     </li>:<></>
                 )
             }
@@ -175,9 +305,10 @@ function Project() {
             {
                 data.map(x=>
                     x.status=="Completed" && x.isDeleted==false?
-                    <li draggable onDragEnd={(e)=>dragToDelete(x._id)} >
+                    <li draggable onDragEnd={(e)=>dragToDelete(x._id,x.assignTo)} >
                         <h4>{x.name}</h4><br/>
-                        {/* <p>{x.description}</p> */}
+                        <p >assignTo: {x.assignTo}</p>
+                        <a style={{cursor: "pointer"}} onClick={()=>{getComments(x._id,x.name)}}><p>comments: {x.totalComment}</p></a>
                     </li>:<></>
                 )
             }
@@ -185,7 +316,69 @@ function Project() {
         
     </div>
         
-    </div>
+        </div>:
+        <>
+        <h3>Task: {taskName}</h3>
+        <section style={{backgroundColor: "#eee"}}>
+            <div class="container my-5 py-5">
+                <div class="row d-flex justify-content-center">
+                <div class="col-md-12 col-lg-10 col-xl-8">
+                    <div class="card">
+                    {
+                        
+                        comments.map(x=>
+                            
+                            !x.isDeleted?
+                            
+                            <div class="card-body" style={{borderStyle:"outset"}}>
+                                <div class="d-flex flex-start align-items-center">
+                                
+                                <div style={{display:"flex"}}>
+                                    <h5 class="fw-bold text-primary mb-1">{x.userId.name}</h5>
+                                    <p style={{marginTop:"10px", marginLeft:"5px"}} class="text-muted small mb-0">{x.dateTime}</p>
+    
+                                </div>
+                                </div>
+                
+                                <p class="mt-3 mb-4 pb-2">{x.message}</p>
+                                <div class="small d-flex justify-content-start" style={{display :"flex"}}>
+                                    {
+                                        userId==x.userId._id?
+                                        <a onClick={()=>deleteComment(x._id)} class="d-flex align-items-center me-3"  style={{cursor: "pointer"}}>
+                                            <p class="mb-0" style={{marginRight:10}}>delete</p>
+                                        </a>:<></>
+                                    }
+
+                                </div>
+                            </div>
+
+                                
+                            :<></>
+    
+                            
+                        
+                        )
+                    }
+                    <div class="card-footer py-3 border-0" style={{backgroundColor: "#f8f9fa"}}>
+                        <div class="d-flex flex-start w-100">
+
+                        <div class="form-outline w-100">
+                            <textarea class="form-control" placeholder="type your comment...." rows="4" style={{background: "#fff"}} onChange={(e)=>{setMsg(e.target.value)}}/>
+                            <label class="form-label" for="textAreaExample">Comment</label>
+                        </div>
+                        </div>
+                        <div class="float-end mt-2 pt-1">
+                        <button type="button" class="btn btn-primary btn-sm" onClick={()=>addComment()}>Post comment</button>
+                        <button type="button" class="btn btn-outline-primary btn-sm" onClick={()=>setShowComments(false)}>Cancel</button>
+                        </div>
+                    </div>
+                    </div>
+                </div>
+                </div>
+            </div>
+        </section>
+        </>
+     }   
     </>
   )
 }
